requirements.txt
discord.py[voice]
spotipy
cryptography
Flask
